feat(campaign): Enhance campaign evaluation logic and add new POS simulation link

+++ src/main/java/com/opensource/campaignengine/dto/CartItemDTO.java
@@ -6,7 +6,7 @@ import java.math.BigDecimal;
 @Data // Lombok'un getter, setter, toString vb. metotları oluşturması için
 public class CartItemDTO {
     private String productId; // Ürünün barkodu veya benzersiz kodu
-    private int quantity;     // Sepetteki adedi
+    private BigDecimal  quantity;     // Sepetteki adedi
     private BigDecimal unitPrice; // Ürünün birim fiyatı
 }
